adds timer on purging, now purges at most every 60 seconds. simplifies purging code logic

+++ sensor_fusion_service/main.py
@@ -75,10 +75,14 @@ def main(db_path: str, debug: bool = False):
     ############################### Main Service Loop ###############################
     if debug:
         loop_counter = 0
+
+    loop_time = time.time()
     while 1:
         ########### Purge DB if required ###########
-        # TODO: Modify to not be every loop, this can be done much less frequently
-        db.purge()
+        # Check every minute if the db needs to be purged
+        if loop_time + 60 < time.time():
+            db.purge()
+            loop_time = time.time()
 
         ########### Check for enough Data for Processing ###########
         ### Find where to start raw index
